stylus_api/requirements.txt
asgiref==3.3.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.6.0
Django==3.1.4
django-allauth==0.44.0
django-cors-headers==3.5.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
idna==2.10
oauthlib==3.1.0
Pillow==8.0.1
pycparser==2.20
PyJWT==1.7.1
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
stripe==2.55.1
urllib3==1.26.2
